requirements.txt
python-dotenv==0.21.1
requests==2.28.2
selenium==4.8.0